// src/app/models/account.model.ts
export interface AccountDetails {
  accountId: string;
  balance: number;
  currentPage: number;
  totalPages: number;
  pageSize: number;
  operationDTOs: AccountOperation[];
}

export interface AccountOperation {
  id: number;
  dateOperation: Date;
  amount: number;
  type: string;
  description: string;
}